creating label file + adjusting label so it is in front of polygon borders
