+++ lib/src/layer/polygon_layer.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 
 import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map/src/layer/label.dart';
 import 'package:flutter_map/src/map/map.dart';
 import 'package:latlong2/latlong.dart' hide Path; // conflict with Path from UI
 
@@ -226,42 +227,6 @@ class PolygonPainter extends CustomPainter {
     }
   }
 
-  void _paintText(Canvas canvas, List<Offset> points) {
-    var maxDx = 0.0;
-    var minDx = double.infinity;
-    var dx = points.map((e) => e.dx).toList().fold<double>(0.0,
-            (previousValue, element) {
-          maxDx = max(maxDx, element);
-          minDx = min(minDx, element);
-          return previousValue + element;
-        }) /
-        points.length;
-    var dy = points.map((e) => e.dy).toList().fold<double>(
-            0.0, (previousValue, element) => previousValue + element) /
-        points.length;
-
-    final textSpan =
-        TextSpan(text: polygonOpt.label, style: polygonOpt.labelStyle);
-    final textPainter = TextPainter(
-      text: textSpan,
-      textAlign: TextAlign.center,
-      textDirection: TextDirection.ltr,
-      maxLines: 1,
-    );
-    if (dx > 0) {
-      textPainter.layout(minWidth: 0, maxWidth: double.infinity);
-      dx = dx - textPainter.width / 2;
-      dy = dy - textPainter.height / 2;
-
-      if (maxDx - minDx > textPainter.width) {
-        textPainter.paint(
-          canvas,
-          Offset(dx, dy),
-        );
-      }
-    }
-  }
-
   void _paintPolygon(Canvas canvas, Rect rect) {
     final paint = Paint();
 
@@ -297,11 +262,16 @@ class PolygonPainter extends CustomPainter {
       path.addPolygon(polygonOpt.offsets, true);
       canvas.drawPath(path, paint);
 
-      if (polygonOpt.label != null) {
-        _paintText(canvas, polygonOpt.offsets);
-      }
-
       _paintBorder(canvas);
+
+      if (polygonOpt.label != null) {
+        Label.paintText(
+          canvas,
+          polygonOpt.offsets,
+          polygonOpt.label,
+          polygonOpt.labelStyle,
+        );
+      }
     }
   }
